+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file, following t
 
 ## [Unreleased]
 
+## [0.6.0] - 2024-03-13
+
 - Fixed tooltip placement bug (not moving when scrolling page)
 - Pinned tooltip moves on zoom/pan/resize
 - Some work on efficient downscaling and anti-aliasing (not ready)